+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- Adding the ability to specify json as an output format when listing versions of an app available
+- Update dependencies
+
 ## [0.17.0] - 2024-10-11
 - Added support for F-Droid entry point specification and new index versions
 - Update dependencies